pyproject.toml
[tool.poetry]
name = "spin"
version = "0.1.0-dev"
description = "A tool to build, start and stop virtual machines"
authors = ['Martín Paradiso <martinparadiso@outlook.com>']
readme = 'README.md'
packages = [{ include = 'spin' }]

[tool.poetry.scripts]
spin = 'spin.cli:run'

[tool.poetry.dependencies]
python = '^3.8'

PyYAML = "^6.0"
typing-extensions = { version = "^4.4.0", python = "<3.10" }
libvirt-python = { version = "^9.0", optional = true }
guestfs = { url = "https://download.libguestfs.org/python/guestfs-1.40.2.tar.gz", optional = true}
uniseg = "^0.7.2"
pyxdg = "^0.28"
tabulate = "^0.9.0"
tomli = {version = "^2.0.1", python = "<3.11"}
pydantic = "^1.10.5"
jinja2 = "^3.1.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
docker = "^6.0.1"
coverage = "^7.0.0"
pytest-cov = "^4.0.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
Sphinx = "^6.0.0"
sphinx-rtd-theme = "^1.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0"
mypy = "^1.1"
pylint = "^2.16.2"
jsonschema = "^4.17.3"

[tool.poetry.extras]
libvirt = ["libvirt-python", "guestfs"]

[tool.pytest.ini_options]
markers = [
    """requires_backend: tests requiring a functional backend to run.""",

    """slow: tests that are slow; from ~0.01s to a couple of seconds""",

    """super_slow: tests that are *really* slow; maybe even several minutes.""",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
include = [ "spin", "tests" ]

[tool.isort]
profile = 'black'

# Code style
[tool.flake8]
max-line-length = 88 # Per black

[tool.pylint."MESSAGES CONTROL"]
disable = [ 
    "wrong-import-position", 
    "invalid-name", 
    "unused-argument", 

    # Pyright suggests explicitly adding ellipsis on abstract methods with
    # docstring, to indicate the intentional lack of implementation.
    "unnecessary-ellipsis",

    # Linter complains about classes with less than 2 'public' methods:
    # this is not relevant to python since we have for instance dataclasses
    # or classes with reduced responsibilities.
    'too-few-public-methods',
]
extension-pkg-allow-list = "pydantic"

[tool.mypy]
disable_error_code = ['method-assign']
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["xdg", 'libvirt', 'docker']
ignore_missing_imports = true

